refactor: split for folders in domain in file service
