+++ backend/src/FileService/FileService.Core/Files/GetDownloadUrlEndpoint.cs
@@ -1,4 +1,4 @@
-﻿using FileService.Domain;
+﻿using FileService.Domain.ValueObjects;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
