+++ backend/src/FileService/FileService.Core/Multipart/CompleteMultipartUpload.cs
@@ -1,7 +1,7 @@
 ﻿using CSharpFunctionalExtensions;
 using FileService.Contracts;
 using FileService.Core.Files;
-using FileService.Domain;
+using FileService.Domain.Entities.MediaAssetEntity;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
