+++ backend/src/FileService/FileService.Domain/Enums/AssetTypeEnum/AssetType.cs
@@ -1,4 +1,4 @@
-﻿namespace FileService.Domain;
+﻿namespace FileService.Domain.Enums.AssetTypeEnum;
 
 /// <summary>
 /// Бизнес тип ассета
