+++ backend/src/FileService/FileService.Domain/Enums/MediaType.cs
@@ -1,4 +1,4 @@
-﻿namespace FileService.Domain;
+﻿namespace FileService.Domain.Enums;
 
 /// <summary>
 /// Категории MIME-type
