--- backend/src/FileService/FileService.Domain/IMediaAssetFactory.cs
@@ -1,11 +1,0 @@
-using CSharpFunctionalExtensions;
-using Shared;
-
-namespace FileService.Domain;
-
-public interface IMediaAssetFactory
-{
-    Result<VideoAsset, Error> CreateVideoForUpload(MediaData mediaData, MediaOwner owner);
-
-    Result<PreviewAsset, Error> CreatePreviewForUpload(MediaData mediaData, MediaOwner owner);
-}
