--- backend/src/FileService/FileService.Domain/MediaAssetFactory.cs
@@ -1,17 +1,0 @@
-using CSharpFunctionalExtensions;
-using Shared;
-
-namespace FileService.Domain;
-
-public sealed class MediaAssetFactory : IMediaAssetFactory
-{
-    public Result<VideoAsset, Error> CreateVideoForUpload(MediaData mediaData, MediaOwner owner)
-    {
-        return VideoAsset.CreateForUpload(Guid.NewGuid(), mediaData);
-    }
-
-    public Result<PreviewAsset, Error> CreatePreviewForUpload(MediaData mediaData, MediaOwner owner)
-    {
-        return PreviewAsset.CreateForUpload(Guid.NewGuid(), mediaData);
-    }
-}
