+++ backend/src/FileService/FileService.Domain/ValueObjects/StorageKey.cs
@@ -1,7 +1,7 @@
 ﻿using CSharpFunctionalExtensions;
 using Shared;
 
-namespace FileService.Domain;
+namespace FileService.Domain.ValueObjects;
 
 /// <summary>
 /// Bucket = "videos"
